Don't blow up art mailer if there is no current open studios
problem
-----

with the recent mod to have scheduled open studios that go away,
we didn't hit all the cases where we assume the open studios event is present.

one showed up while testing something in acceptance.  Adding art generates an email which is counting on open studios event.

solution
------

Add some guards and fix the email.

+++ app/services/open_studios_participation_service.rb
@@ -1,12 +1,16 @@
 # manage caching the list of open studios events
 class OpenStudiosParticipationService
   def self.participate(artist, open_studios_event)
+    return unless open_studios_event
+
     OpenStudiosParticipant.transaction do
       artist.open_studios_participants.create(open_studios_event:) unless artist.open_studios_participants.find_by(open_studios_event:)
     end
   end
 
   def self.refrain(artist, open_studios_event)
+    return unless open_studios_event
+
     artist.open_studios_participants.where(open_studios_event:).destroy_all
   end
 end
